Limit tiles sent to player based on location
During this commit I played around with increasing the map size and checking performance. I was able to create maps as large as 1000x1000 when increasing memory allocation on the server, but it was taking 23-25 seconds to tick the game world. I noticed that the majority of that time seemed to be in sending state to the player, so I looked at how I could reduce that. Now, when sending state, the game will create an instance of a special player state class, and instead of sending all of the tiles in the world to each player, will include only tiles relevant to that player in the player state object before sending it. Relevant tiles are chosen by proximity to the player and are retrieved from the Map class.